fixing ctro
